api/requirements.txt
Flask==2.0.2
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
flask-restful 
redis
flask-cors
werkzeug
numpy
gunicorn==20.0.4
pandas
snps[ezancestry]